Avoid endless recursion, property access and Dictionary keys errors.
Avoid recursing into objects in the System.Reflection namespace, since
these objects create new object instances on the fly, causing endless recursion.

Catch and dump additional exceptions that can occur when trying to retrieve
the value of properties through refelection.

Some objects that the ObjectDumper may encounter cannot be used as a dictionary
key, even if it is not a value type. Therefore, use the more robust
ObjectIDGenerator to track which objects that have already been seen and
assign them numeric identifiers.